cloud-proxy/requirements.txt
Flask
google-cloud-secret-manager
requests
